Se agrego dao COMPRAPRODUCTO y se agrego a la BD

+++ Natura/src/DatabaseDAO/productoDAO.java
@@ -2,5 +2,6 @@ package DatabaseDAO;
 
 import Clases.Producto;
 
-public interface productoDAO extends DAO<Producto,Short>{
+public interface productoDAO extends DAO<Producto,Integer>{
+    
 }
